# modules/swagger-jaxrs-client-kek/swagger-inputs/swagger.yaml
swagger: "2.0"
info:
  description: "BaaS for delivery services like Glovo, Uber or even for regular mail. Also, a simple web front-end should be provided as an example of a typical consumer's app."
  version: "1.0.0"
  title: "Kinda Express King"
#  contact:
#    email: "apiteam@softserveinc.com"
# host: "localhost:8080"
...
paths:
  /users:
    get:
      tags:
        - "users"
      summary: "Searches for users"
      description: "Finds all users"
      operationId: "getUserList"
      produces:
        - "application/vnd.softserve.userList+json"
      responses:
        200:
          description: "List of users"
          schema:
            $ref: "#/definitions/userList"
    post:
      tags:
        - "users"
      summary: "Creates a new user"
      operationId: "addUser"
      consumes:
        - "application/vnd.softserve.user+json"
      produces:
        - "application/vnd.softserve.user+json"
        - "application/vnd.softserve.errorList+json"
      parameters:
        - name: "user"
          in: "body"
          description: "The user to create"
          schema:
            $ref: "#/definitions/user"
      responses:
        200:
          description: "The newly created user object"
          schema:
            $ref: "#/definitions/user"
        400:
          description: "Fields validation failed"
          schema:
            $ref: "#/definitions/errorList"
  /users/{guid}:
    get:
      tags:
        - "users"
      summary: "Finds the specific user"
      operationId: "getUser"
      produces:
        - "application/vnd.softserve.user+json"
      parameters:
        - name: "guid"
          in: "path"
          description: "ID of the specific user"
          required: true
          type: "string"
      responses:
        200:
          description: "The user object"
          schema:
            $ref: "#/definitions/user"
    put:
      tags:
        - "users"
      summary: "Modifies the specific user"
      operationId: "modifyUser"
      consumes:
        - "application/vnd.softserve.user+json"
      produces:
        - "application/vnd.softserve.user+json"
        - "application/vnd.softserve.errorList+json"
      parameters:
        - name: "guid"
          in: "path"
          description: "ID of the specific user"
          required: true
          type: "string"
        - name: "user"
          in: "body"
          description: "The user to modify"
          schema:
            $ref: "#/definitions/user"
      responses:
        200:
          description: "The modified user object"
          schema:
            $ref: "#/definitions/user"
        400:
          description: "Fields validation failed"
          schema:
            $ref: "#/definitions/errorList"
    delete:
      tags:
        - "users"
      summary: "Deletes the specific user"
      operationId: "deleteUser"
      parameters:
        - name: "guid"
          in: "path"
          description: "ID of the specific user"
          required: true
          type: "string"
      responses:
        200:
          description: "Successful operation"
  /users/{guid}/addresses:
    get:
      tags:
        - "users"
      summary: "Finds addressess of the specific user"
      operationId: "getUserAddresses"
      produces:
        - "application/vnd.softserve.addressList+json"
      parameters:
        - name: "guid"
          in: "path"
          description: "ID of the specific user"
          required: true
          type: "string"
      responses:
        200:
          description: "List of user addresses"
          schema:
            $ref: "#/definitions/addressList"
    post:
      tags:
        - "users"
      summary: "Adds a new addresses"
      operationId: "addUserAddresses"
      consumes:
        - "application/vnd.softserve.addressList+json"
      produces:
        - "application/vnd.softserve.addressList+json"
        - "application/vnd.softserve.errorList+json"
      parameters:
        - name: "guid"
          in: "path"
          description: "ID of the specific user"
          required: true
          type: "string"
        - name: "addresses"
          in: "body"
          description: "List of addresses"
          schema:
            $ref: "#/definitions/addressList"
      responses:
        200:
          description: "List of added user addresses"
          schema:
            $ref: "#/definitions/addressList"
        400:
          description: "Fields validation failed"
          schema:
            $ref: "#/definitions/errorList"
  /users/{guid}/addresses/{addrguid}:
    get:
      tags:
        - "users"
      summary: "Finds addressess of the specific user"
      operationId: "getUserAddress"
      produces:
        - "application/vnd.softserve.address+json"
      parameters:
        - name: "guid"
          in: "path"
          description: "ID of the specific user"
          required: true
          type: "string"
        - name: "addrguid"
          in: "path"
          description: "ID of the specific address"
          required: true
          type: "string"
      responses:
        200:
          description: "Specific user address"
          schema:
            $ref: "#/definitions/address"
    put:
      tags:
        - "users"
      summary: "Modifies the specific user address"
      operationId: "modifyUserAddress"
      consumes:
        - "application/vnd.softserve.address+json"
      produces:
        - "application/vnd.softserve.address+json"
        - "application/vnd.softserve.errorList+json"
      parameters:
        - name: "guid"
          in: "path"
          description: "ID of the specific user"
          required: true
          type: "string"
        - name: "addrguid"
          in: "path"
          description: "ID of the specific address"
          required: true
          type: "string"
        - name: "address"
          in: "body"
          description: "The address to modify"
          schema:
            $ref: "#/definitions/address"
      responses:
        200:
          description: "The modified address object"
          schema:
            $ref: "#/definitions/address"
        400:
          description: "Fields validation failed"
          schema:
            $ref: "#/definitions/errorList"
    delete:
      tags:
        - "users"
      summary: "Deletes the specific user address"
      operationId: "deleteUserAddress"
      parameters:
        - name: "guid"
          in: "path"
          description: "ID of the specific user"
          required: true
          type: "string"
        - name: "addrguid"
          in: "path"
          description: "ID of the specific address"
          required: true
          type: "string"
      responses:
        200:
          description: "Successful operation"
...
definitions:
  userList:
    type: "object"
    properties:
      userList:
        type: "array"
        items:
          $ref: "#/definitions/user"
  user:
    type: "object"
    properties:
      guid:
        type: "string"
      name:
        type: "string"
      nickname:
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
      details:
        $ref: "#/definitions/userDetails"
  userDetails:
    type: "object"
    properties:
      payload:
        type: "string"
      imageURL:
        type: "string"
...
  addressList:
    type: "object"
    properties:
      addressList:
        type: "array"
        items:
          $ref: "#/definitions/address"
  address:
    type: "object"
    properties:
      guid:
        type: "string"
      alias:
        type: "string"
      address:
        type: "string"
      notes:
        type: "string"
  errorList:
    type: "object"
    properties:
      errorlList:
        type: "array"
        items:
          $ref: "#/definitions/error"
  error:
    type: "object"
    properties:
      error:
        type: "string"
      details:
        type: "string"

# Added by API Auto Mocking Plugin
host: virtserver.swaggerhub.com
basePath: /burnkr/kek/1.0.0
schemes:
  - https